nav manu move and reintro auth redirector
